backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
gunicorn
django-cors-headers
python-dotenv
drf-spectacular
Pillow
